fix: fix types pathes
