+++ src/types.ts
@@ -6,3 +6,12 @@ export interface User {
 }
 
 export type db = User[];
+
+export type MethodHandlers = {
+  GET: () => Promise<void>;
+  POST: () => Promise<void>;
+  PUT: () => Promise<void>;
+  DELETE: () => Promise<void>;
+};
+
+export type Method = keyof MethodHandlers;
